requirement.txt
sentence-transformers
faiss-cpu
pandas
scikit-learn
numpy
PyPDF2

